Federation Association Club Test Green!!!!! OK (46 tests, 365 assertions)

+++ resources/views/clubs/index.blade.php
@@ -10,13 +10,13 @@
     <div class="container-fluid">
 
         @if (sizeof($clubs)==0)
-            {{--            @include('layouts.noFederations')--}}
+            @include('layouts.noClubs')
         @else
 
             @can('create', new App\Club)
                 <span class="pl-10 pull-right">
-                    <a href="{!!   URL::action('ClubController@create') !!}"
-                       class="btn btn-primary btn-xs "><b><i class="icon-plus22 mr-5"></i></b>
+                    <a id="addClub" href="{!!   URL::action('ClubController@create') !!}"
+                       class="btn btn-primary btn-xs "><b><i class="icon-plus22 mr-5" ></i></b>
                         @lang('core.addModel', ['currentModelName' => $currentModelName])
                     </a>
                 </span>
@@ -38,20 +38,28 @@
                 </thead>
                 @foreach($clubs as $club)
                     <tr>
-                        <td><a
-                                    href="{!!   URL::action('ClubController@edit',  $club->id) !!}">{{ $club->name }}</a>
+                        <td>@can('edit',$club)
+                                <a href="{!!   URL::action('ClubController@edit',  $club->id) !!}">{{ $club->name }}</a>
+                            @else
+                                {{ $club->name }}
+                            @endcan
                         </td>
-                        <td>{{ $club->association !=null ? $club->association->federation->name : "-"}}</td>
+                        <td>{{ $club->association !=null && $club->association->federation != null ? $club->association->federation->name : "-"}}</td>
                         <td>{{ $club->association !=null ? $club->association->name : "-"}}</td>
                         <td align="center">{{ $club->president !=null ? $club->president->name : " - " }}</td>
                         <td align="center">{{ $club->president != null ? $club->president->email : " - "}}</td>
                         <td align="center">{{ $club->address }}</td>
                         <td align="center">{{ $club->phone }}</td>
                         <td align="center">
-                            <a href="{{URL::action('ClubController@edit', $club->id)}}"><i class="icon icon-pencil7"></i></a>
-                            {!! Form::open(['method' => 'DELETE', 'id' => 'formDeleteClub', 'action' => ['ClubController@destroy', $club->id], 'style'=>"display: inline-block"]) !!}
-                            {!! Form::button( '<i class="glyphicon glyphicon-trash"></i>', ['type' => 'submit','class' => 'btn text-warning-600 btn-flat btnDeleteClub', 'id'=>'delete_'.$club->id, 'data-id' => $club->id ] ) !!}
-                            {!! Form::close() !!}
+                            @can('edit',$club)
+                                <a href="{{URL::action('ClubController@edit', $club->id)}}"><i
+                                            class="icon icon-pencil7"></i></a>
+                            @endcan
+                            @can('delete',$club)
+                                {!! Form::open(['method' => 'DELETE', 'id' => 'formDeleteClub', 'action' => ['ClubController@destroy', $club->id], 'style'=>"display: inline-block"]) !!}
+                                {!! Form::button( '<i class="glyphicon glyphicon-trash"></i>', ['type' => 'submit','class' => 'btn text-warning-600 btn-flat btnDeleteClub', 'id'=>'delete_'.$club->id, 'data-id' => $club->id ] ) !!}
+                                {!! Form::close() !!}
+                            @endcan
                         </td>
 
                     </tr>
